[REFACT] Moved resolving hook targets to: HookTargetResolver
